+++ crates/gpui/src/platform/mac/window.rs
@@ -2342,10 +2342,11 @@ extern "C" fn do_command_by_selector(this: &Object, _: Sel, _: Sel) {
     drop(lock);
 
     if let Some((keystroke, mut callback)) = keystroke.zip(event_callback.as_mut()) {
+        let prefer_character_input = keystroke.modifiers.alt && !keystroke.is_ascii();
         let handled = (callback)(PlatformInput::KeyDown(KeyDownEvent {
             keystroke,
             is_held: false,
-            prefer_character_input: false,
+            prefer_character_input,
         }));
         state.as_ref().lock().do_command_handled = Some(!handled.propagate);
     }